Attempted Changes
I think I have this right.  But there is an error that comes up.  I was
attempting to make it work with the tabs, but I decided on changing the
format to push buttons and a text browser.  Since I had a previous slot
for the tab, it is still showing up on my error message and I don't know
how to fix it.  If that goes away, then i am fairly confident that this
will work.